+++ desc/objectives/_coils.py
@@ -965,10 +965,10 @@ class CoilArclengthVariance(_CoilObjective):
 
     This objective is meant to combat any issues corresponding to non-uniqueness of
     the representation of a curve, in that the same physical curve can be represented
-    by different parametrizations by changing the curve parameter [1]_. Note that this
-    objective has no effect for ``FourierRZCoil`` and ``FourierPlanarCoil`` which have
-    a single unique parameterization (the objective will always return 0 for these
-    types).
+    by different parametrizations by changing the curve parameter [1]_.
+    Note that this objective has no effect for ``FourierRZCoil``, ``FourierPlanarCoil``,
+    and ``FourierXYCoil`` which have a single unique parameterization
+    (the objective will always return 0 for these types).
 
     References
     ----------